// packages/timeback-clients/src/oneroster/generated/operation-specs.ts
export const operationSpecs = defineOperationSpecs({
  getAllAssessmentLineItems: {
// ...
    response: z
      .object({
        assessmentLineItems: z.array(
          z
            .object({
              sourcedId: z.string().min(1).optional(),
              status: z.enum(["active", "tobedeleted"]),
              dateLastModified: z
                .string()
                .datetime({ offset: true })
                .optional(),
              metadata: z.object({}).partial().passthrough().nullish(),
              title: z.string(),
              description: z.union([z.string(), z.null()]).optional(),
              class: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
              parentAssessmentLineItem: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
              scoreScale: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
              resultValueMin: z.union([z.number(), z.null()]).optional(),
              resultValueMax: z.union([z.number(), z.null()]).optional(),
              component: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
              componentResource: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
              learningObjectiveSet: schemas.LearningObjectiveSet.optional(),
              course: z
                .union([
                  z.object({ sourcedId: z.string() }).passthrough(),
                  z.null(),
                ])
                .optional(),
            })
            .passthrough()
        ),
        totalCount: z.number(),
        pageCount: z.number(),
        pageNumber: z.number(),
        offset: z.number(),
        limit: z.number(),
      })
      .passthrough(),
    errors: [
      {
        status: 400,
        description: "Bad Request",
        schema: schemas.BadRequestResponse,
      },
      {
        status: 401,
        description: "Unauthorized",
        schema: schemas.UnauthorizedRequestResponse,
      },
      {
        status: 403,
        description: "Forbidden",
        schema: schemas.ForbiddenResponse,
      },
      {
        status: 404,
        description: "Not Found",
        schema: schemas.NotFoundResponse,
      },
      {
        status: 422,
        description: "Unprocessable Entity / Validation Error",
        schema: schemas.UnprocessableEntityResponse,
      },
      {
        status: 429,
        description: "Too Many Requests",
        schema: schemas.TooManyRequestsResponse,
      },
      {
        status: 500,
        description: "Internal Server Error",
        schema: schemas.InternalServerErrorResponse,
      },
    ] as const,
  },
// ...
} as const satisfies OperationSpecMap);

export type OneRosterOperationSpecs = typeof operationSpecs;
export type OneRosterOperationAlias = keyof OneRosterOperationSpecs;